backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
coverage==7.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pyasn1==0.4.8
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
rsa==4.9
ruff==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
